Use lighter-weight check for origin classloader of invokers.
Instead of using somewhat expensive getMethod calls, this uses
a simple comparison of the found class's superclass an expected
class from the current classloader context. The effect should be
the same...that is, if an invoker comes from a higher classloader
with mismatched JRuby classes, it will fail this check just as it
failed the getMethod check.

See JRUBY-5038 and related commits.