Add RubyString.newUTF16String and newUTF8String.
These new methods explicitly produce UTF-16 and UTF-8 Ruby strings
from incoming Java strings. The old newUnicodeString method will
now choose UTF-8 or UTF-16 based on whether the default internal
encoding rather than always doing UTF-8.

This allows a user to specify that the Ruby default internal
encoding should be UTF-16, and we will start encoding all Java
strings accordingly. This may not be a common way to run, but for
situations requiring the fastest-possible Java string handling, it
is an interesting option.

This also seems like "the right thing to do" if a user sets UTF-16
as default internal encoding, so it has a logical justification
outside of performance too.