Patch around JI specs that break with new Java 8 collections.
Specifically, Java 8 adds #sort to List (sorts in place given a
comparator, so our zero-arg calls and return value expectations
fail), #replace to Map (replaces a given key with a given value),
and #merge to Map (merges a given key's value with an incoming
value). This causes our monkey-patched versions to not be visible
and a few specs to break.