added self-first strategy to JRubyClassLoader with ITs for jruby-main and jruby-complete
the motivation for the patch is the following:

currently when there is for example bouncy-castle jars in the classpath or in the parent
classloader of the JRubyClassloader then you can install any bouncy-castle-java gem as you
wish but it will have no effect since the classloaders first look at the parent classloader.
the same holds true for any vendored jar which comes with a gem. this patch simple first looks
at it self when loading a class before it decends to the parent classloader to load it.