Use highest compat when binding methods using it in JRubyMethod.
This is necessary to maintain some backward compat with libraries
that needed to separate compat-specific behavior for a given
method. We no longer use compat in JRuby proper, but in order to
let those libraries work this commit modifies method binding to
choose the highest compat of all defined methods when compat is
being used. This should be correct for all cases, since no version
of JRuby was released with official support for anything higher
than 1.9, meaning the highest annotation we should see in the
wild would be 1.9 compat.

Fixes #1194.