Use frame class from block, not from context. Fixes #1318.
There's no frame from which to get the class, so we can't get it
from ThreadContext here. Get it from the proc.

Note that there's a more difficult issue here in that there's also
no frame/backtrace/etc on the ThreadContext for invoking a
TracePoint interested in these thread events. Fixing that might
mean making a frame be available at this point, making this fix
unnecessary...but either way the class reference used in this fix
will still be correct.