Implement deduplication of literal hash string keys.
Because we know that string keys are immediately duplicated and
frozen we can use our freeze-dup cache of strings in the compiler
whenever we see a literal hash with literal string keys. This
avoids the overhead of duplicating and freezing literal strings
repeatedly, for a roughly 4x performance speedup on a 5-element
literal hash with literal string keys. In addition, this commit
modifies the normal freeze-dup process in Hash to use the dedup
cache added in 926ca89.

This commit may introduce contention since the process of freeze-
duping string keys now must pass through the synchronized dedup
logic from 926ca89. We may need to reexamine that logic, or not
use it in the hash key string insertion freeze-duping.