Temporary fix for cost of Hash#[] dynamically calling #default.
This fix adds "default" to the list of methods we build up a
per-runtime set of call sites for. I consider this a temporary
fix because that set of sites is a matrix of class count * method
count, so adding default adds another entire row/column when it
will only be used from one class.

It's clear we need some better logic for doing call site caching
for dynamic calls in Java code, but this will work for the moment.