[IR] Bug fix in handlePropagatedBreak
Commit 7a97603f converted IRBreakJump to Unrescuable.

For some reason, this broke exit from irb on hitting ^D where an
unrescuable ruby exception was being processed by handlePropagatedBreak.

This was a legitimate bug in handlePropagatedBreak which seems to
have been exposed by the fix. This patch fixed that runtime helper
to correctly handle ruby exceptions that go through it.

With this fix, handlePropagatedBreak mirrors the structure of
handleNonLocalReturn (which makes more sense).