Fix class/module definition and ensure JVM method uniqueness.
Class/module definition had not been updated for changes in the
signature of CompiledIRMethod.<init>; specifically, they did not
pass an empty String to represent no parameter names. The handle
for the class/module body was also being generated with the wrong
signature.

Uniqueness for JVM methods was done by line number, which is not
sufficient for when multiple of the same type of constructs
appeared on the same line. I made it a simple incrementing integer
for the script for now.