[Truffle] - Reusing JRuby/RubyString#op_equal code on RubyString#equal.
In order to make the integration between ruby and truffle more smooth we
can reuse the code instead of implement it from scratch. In this
particular case I'm converting the Truffle/RubyString into a
JRuby/RubyString and then call the methods defined in JRuby/RubyString
and then converting the response back to the format expected by truffle.