Switch to use resources to get directory entries.
Current jar Dir.entries implementation doesn't return "." and ".." so to keep this behavior in 1.7,
I've moved the "." and ".." addition logic to RegularFileResource. Unfortunately fixing the jar files to
support such behavior would require and invasive refactor of File/Dir implementations to be able to support
expanding '..' in jar file names.