If the LoadService load path file watcher isn't running, just degrade to the old behavior.
There's no point in constantly doing File#listFiles() calls if we're ultimately tossing the results each time. If the path doesn't contain the file being required, we still hit the file system once (loadFiles call), but if does contain, we hit the file system twice (loadFiles then canRead). So, the loadFiles call only makes sense if we can cache and we can only cache if we can invalidate.