Hook up a mixed-mode for the IR runtime, to aid JIT testing.
Like the old AST-based runtime, we have both an interpreter and
a compiler. Up to now, it has only been possible to run the IR
runtime in *either* the interpreter or the still-in-development
JIT. This patch adds a -X+JIR mode that will lazily JIT IR method
bodies using the same trivial heuristic (call count) as the AST
runtime.

This is a major step toward switching to IR runtime being on all
the time.