Fix incorrect assignment of chFile to null.
This regressed us on some specs because of logic specific to
FileChannel starting to kick in where it did not before. Basically
anything that was checking for chFile before always failed and
used alternative logic, so that logic was not getting tested.

I'm committing this because it's obviously wrong and I need to
back up on passing specs to fix them again.