Move toward better metadata in native core methods.
* Fully-specify methods that read/write frame or scope fields.
* Attach additional "caller" CallConfig to methods for later use.
* Eliminate static lists of scope/frame consumers (full field
  specification means those methods will get picked up on boot.

This change moves us toward having full framing information in the
DynamicMethod object associated with a given native method. This
will allow us to do a better job of late-framed calls (don't
frame or scope until we call a method that needs it), alias
warnings (only warn when aliasing a method that has framing
requirements), and ultimately better frame elimination in the new
IR runtime.