Inline CRC32Ext
Compared to the inlined version of Adler32, this one unfortunately still
relies on reflection for adjusting the starting checksum. This is due
to the fact that crc32_combine is much more expensive, and completely
dominates the runtime up to quite long byte lists.

However, in the event that the reflection trick stops working, we now
fall back to the slower approach, ensuring compatibility with an
unlikely future Java release where the internal CRC32 field name is
changed.

Compared to CRC32Ext, however, neither the reflection nor the slow path
is used when the starting checksum matches the default.