* stay within the classpath: protocol
* unsure we pass on the Url with its URLStreamHandler provided by the classloader directly to the JRubyClassloader

remark on the URL:
currently there are plenty of issues where those embedded jars from jruby-stdlib
can not be loaded. AND if somehting goes wrong adding that URL to the JRubyClassloader
there will be no Exception so the actual error pops as some constant not found or
some import_java does not work. turning on debug.loadService often shows that jar could
be found and still it is not working. URL -> toString -> new URL( String ) is not reliable since
you always can construct URL objects with "custom" URLStreamHandler. the idea here to
just take whatever URL the classloader produces and pass it on the JRubyClassloader.