Use appropriate isolated scope for ScriptingContainer here.
The default constructor uses SINGLETON scope, which shares a Ruby
instance with the VM running the spec. When the container goes
out of scope and eventually GCs, it also tears itself down. This
in turn causes stdio streams to be marked as closed (though not
actually closed), so the parent runtime can no longer use them.

This is not really a bug, since container teardown should trigger
runtime teardown, and runtime teardown should flush and terminate
all IO streams. However, the old IO logic acted more aggressively
to preserve stdio streams, exposing this flaw in the test.

It may be we need to restore that preservation logic before
release, but for the moment I'm just modifying the test to use the
right scope.