Use MRI lib for open3, deprecate our hidden impls.
We provided IO.popen3 and popen4 previously because open3.rb used
forking. Now that it uses spawn, which is fully supported in 9k,
we can go back to the regular open3.rb. This provides better
compatibility with MRI.

popen4 will be removed. It was never officially available for
users, and it can be achieved by using Open3 and getting the pid
off the IO objects.

The internal implementations of popen3/4 also tried to unwrap the
subprocess's streams from java.lang.Process. This is broken on
Java 7+ because Process logic now eagerly attempts to drain the
child streams in a separate thread, making the unwrapped inner
stream/channel useless. Because 9k currently only supports Java
7+, and because the internal popen3/4 are now deprecated and
unused, I have removed this wrapping.