Ensure blocking tasks have been assigned before going to sleep.
Setting sleep before the blocking task is present can lead to a
separate wakeup thread attempting to unblock the thread before an
unblocker has been set. Ultimately this means we see a thread stay
asleep and the wakeup request is lost.

Accompanying test hung quickly without this change and does not
appear to hang with it.