Remove tesla.io from repos because it's timing out.
repository.tesla.io is used in pluginRepositories to get the Tesla
plugins, but it was also in our regular repository list. I'm not
sure why.

Today, that address is timing out, which has made local builds
impossible. I commented it out in this commit and clean builds
succeed.

@mkristian Do you know why this server is timing out? Do we need
it in the regular repository list?