Deprecate overloaded `setProfile`, prefer `setProfilingMode`
The ScriptingContainer class contains two methods named `setProfile`,
which differ only by their single argument type and do very different
things from one another.  One of the two signatures accepts
an argument of type `ProfilingMode`, which, in other places in the
code, is exposed via more explicitly named accessor methods such as
`getProfilingMode` and `setProfilingMode`.  I'm wondering if maybe
the method in `ScriptingContainer` was simply accidentally mis-named.

This commit deprecates the `setProfile(ProfilingMode)` method, and
introduces a replacement method `setProfilingMode(ProfilingMode)`,
which is more consistent with the rest of the code.