Try JIT twice.
Because the JIT runs additional passes (and may run more in the
future) and because we would like to avoid the interpreter running
code that has passed through those passes, we do a two-pass JIT
here. The first pass tries to compile the code without running the
JIT passes, ensuring it will complete successfully and the code
size will not be too large to load (or so large it never
optimizes in JVM). The second pass then runs our JIT passes and
repeats the compilation, presumably with a successful result.

This should be replaced with cloning the interpreted scope before
doing JIT, for concurrency and overhead reasons (double-JIT cost).