Finally make prepareInstruction clone instrs + labels and not share any state with IRScope.
Problematic that we re-store info back into CFG at this point so debug output matches that
interpreter.  That is ok though since JIT should still be fine with these changes.  It is
not fine if we have debug on and interp is printing out CFG while the JIT is changing the
CFG.  A problem for another day.

Disabled OptimizeDynamicScope for now.  Something with cloning instrs + some closures
makes it break.  It appears not not find lvars in its flattened state.