Switch travis to spec:ruby:fast.
Unfortunately the many specs that re-spawn JRuby are now causing
the build to time out. We have work to do on speeding up boot time
for 9k, but right now we need to get green spec runs on travis.

We can also set up longer test runs on our Jenkins instance. That
may be more appropriate for the long-running suites.