Remove old AST decorators and related specs.
The AST is no longer considered our lowest executable
representation of code, since it requires additional processing
via the IR subsystem. The deleted specs would also need to keep up
with changes to how AST nodes are converted into IR and how to
launch that IR directly into the interpreter. These features are
also not (as far as we know) in use by anyone other than core
JRuby folks (e.g. the jrubyc top-level logic that will have to
be rewritten anyway). For these reasons, I believe it's best that
we remove the functionality and if anyone really needs it we can
spin off a gem that provides the lost functionality in some form.